api: stop using controller-runtime, move dependencymagnet out of package
There is just a small bit of scaffolding we don't really need. Plus,
rather than including the dependency magnet in the api type definition
itself, move it to a separate package.
